+++ CHANGELOG.md
@@ -7,6 +7,7 @@ UNRELEASED
 - Fix typo in russian translation
 - Update browser detection library to 1.4.0
 - Fix width errors where Drop-in was not aligned with other elements on merchant page
+- Add ability to style PayPal button
 
 1.1.0
 ----------